Rename CoopVecAPIExtensions.h to CoopVecAPI.h
